+++ df-chat-enhance/src/edit/DFChatEditor.ts
@@ -1,5 +1,13 @@
+import SETTINGS from "../../../common/Settings";
+
+
+declare namespace marked {
+	function parse(md: string, options: any): string;
+}
+
 export default class DFChatEditor extends FormApplication {
 	private chatMessage: ChatMessage;
+	public static readonly PREF_MARKDOWN = 'edit-markdown';
 
 	/**
 	 * Assign the default options which are supported by the entity edit sheet.
@@ -42,7 +50,11 @@ export default class DFChatEditor extends FormApplication {
 	/** @override */
 	async _updateObject(_event?: any, formData?: any) {
 		let data = formData.content as string;
-		data = data.replace(/\r?\n/gm, '<br/>');
+		if (SETTINGS.get(DFChatEditor.PREF_MARKDOWN)) {
+			data = DFChatEditor.processMarkdown(data)[1];
+		} else {
+			data = data.replace(/\r?\n/gm, '<br/>');
+		}
 		if (data.search(/<p +class="df-edited">/) < 0) {
 			data += `<p class="df-edited">${'DF_CHAT_EDIT.EditedFlag'.localize()}</p>`;
 		}
@@ -54,5 +66,21 @@ export default class DFChatEditor extends FormApplication {
 	close(options?: FormApplication.CloseOptions) {
 		delete (<any>this.chatMessage).chatEditor;
 		return super.close(options);
+	}
+
+	public static processMarkdown(message: string) {
+		const originalMessage = message;
+		message = marked.parse(message, {
+			headerIds: false,
+			breaks: true
+		}).trimEnd();
+		if (message.startsWith('<p>'))
+			message = message.substr(3);
+		if (message.endsWith('</p>'))
+			message = message.substr(0, message.length - 5);
+		const newLine = /(<\/?[ a-z]+>)\n(<\/?[ a-z]+>?)/;
+		while (newLine.test(message))
+			message = message.replace(newLine, '$1$2');
+		return [originalMessage, message];
 	}
 }